# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: AI Trading Platform API
  description: |
    Comprehensive REST API for the AI Trading Platform providing access to trading signals, 
    portfolio management, risk monitoring, and model serving capabilities.
    
    ## Features
    - Multi-exchange trading support (Robinhood, OANDA, Coinbase)
    - AI-powered trading signals with CNN+LSTM models
    - Real-time market data streaming via WebSocket
    - Portfolio optimization and risk management
    - Model interpretability and explainability
    
    ## Authentication
    All endpoints require JWT authentication. Obtain a token via `/auth/login`.
    
    ## Rate Limits
    - Free Tier: 100 requests/hour, 5 signals/day
    - Pro Tier: 1000 requests/hour, 100 signals/day
    - Enterprise: Custom limits
    
  version: 1.0.0
  contact:
    name: AI Trading Platform Support
    email: api-support@ai-trading-platform.com
    url: https://docs.ai-trading-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ai-trading-platform.com
    description: Production server
  - url: https://sandbox-api.ai-trading-platform.com
    description: Sandbox server
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Refresh an existing JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "secure_password"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 3600
        user_id:
          type: string
          format: uuid
        tier:
          type: string
          enum: [free, pro, enterprise]
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        status_code:
          type: integer
        request_id:
          type: string
          format: uuid
        timestamp:
          type: number
...